src/pyproject.toml
[project]
name = "moonclaw-plugins"
version = "0.1.0"
description = "MoonClaw 中的自定义插件"
requires-python = ">=3.11, <4.0"


[tool.poetry.dependencies]
python = ">=3.11, <4.0"
nonebot2 = {extras = ["fastapi"], version = ">=2.3.1, <3.0.0"}

httpx = {extras = ["socks"], version = ">=0.28.1, <0.29.0"}
aiofiles = ">=23.2.1, <24.0.0"
numpy = ">=2.3.1, <3.0.0"
pillow = ">=10.3.0, <13.0.0"
scikit-image = ">=0.23.1, <0.26.0"
jieba = ">=0.42.1, <0.43.0"
sympy = ">=1.12.1, <2.0.0"
openai = ">=1.34.0, <2.0.0"
aiosqlite = ">=0.20.0, <0.21.0"
aiomysql = ">=0.2.0, <0.3.0"
cryptography = ">=44.0.0, <45.0.0"
pyproject-toml = ">=0.0.10, <0.0.11"
matplotlib = ">=3.10.0, <4.0.0"
bilibili-api-python = "^17.4.1"

py-cpuinfo = ">= 9.0.0, < 10.0.0"
html2text = ">=2025.4.15, <2026.0.0"
imagehash = ">=4.3.2, <5.0.0"
markdown = ">=3.7, <4.0.0"
psutil = ">= 7.1.0, < 8.0.0"

nonebot-adapter-onebot = ">=2.4.6, <3.0.0"
nonebot-plugin-htmlrender = ">=0.6.6, <0.7.0"

nonebot-plugin-alconna = ">=0.59.2, <0.60.0"
nonebot-plugin-apscheduler = "^0.5.0"
nonebot-plugin-orm = ">=0.8.0, <0.9.0"
nonebot-plugin-session = ">=0.3.2, <0.4.0"
nonebot-plugin-userinfo = ">=0.2.6, <0.3.0"
nonebot-adapter-qq = {git = "https://github.com/Moonlark-Dev/adapter-qq"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
packages = [
    { include = "nonebot_plugin_access", from = "./plugins" },
    { include = "nonebot_plugin_alconna_extensions", from = "./plugins" },
    { include = "nonebot_plugin_bots", from = "./plugins" },
    { include = "nonebot_plugin_friend_add", from = "./plugins" },
    { include = "nonebot_plugin_larkhelp", from = "./plugins" },
    { include = "nonebot_plugin_larklang", from = "./plugins" },
    { include = "nonebot_plugin_larkuser", from = "./plugins" },
    { include = "nonebot_plugin_larkutils", from = "./plugins" },
    { include = "nonebot_plugin_openai", from = "./plugins" },
    { include = "nonebot_plugin_render", from = "./plugins" },
    { include = "nonebot_plugin_wolfram_alpha", from = "./plugins" },
    { include = "nonebot_plugin_chat", from = "./plugins" },
    { include = "nonebot_plugin_status_report", from = "./plugins" },
    { include = "nonebot_plugin_broadcast", from = "./plugins" },
    { include = "nonebot_plugin_version_manager", from = "./plugins" },
]


[tool.poetry.plugins.nb_scripts]
larkhelp-generate = "nonebot_plugin_larkhelp.__main__:generate_help_markdown"

